+++ src/SFA.DAS.EmployerIncentives.Database/Tables/IncentiveApplicationApprenticeship.sql
@@ -6,7 +6,7 @@
 	[FirstName] NVARCHAR(100) NOT NULL,
 	[LastName] NVARCHAR(100) NOT NULL,
 	[DateOfBirth] DATETIME NOT NULL,
-	[ULN] INT NOT NULL,
+	[ULN] BIGINT NOT NULL,
 	[PlannedStartDate] DATETIME NOT NULL,
 	[ApprenticeshipEmployerTypeOnApproval] NVARCHAR(50),
 	CONSTRAINT FK_IncentiveApplication FOREIGN KEY (IncentiveApplicationId) REFERENCES IncentiveApplication(Id)
